基础能力小优化
