+++ src/main/java/com/kyle/kyleaicodegen/model/dto/user/UserAddRequest.java
@@ -1,19 +1,11 @@
-package com.kyle.kyleaicodegen.model.dto;
+package com.kyle.kyleaicodegen.model.dto.user;
 
-import com.kyle.kyleaicodegen.common.PageRequest;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
 
-@EqualsAndHashCode(callSuper = true)
 @Data
-public class UserQueryRequest extends PageRequest implements Serializable {
-
-    /**
-     * id
-     */
-    private Long id;
+public class UserAddRequest implements Serializable {
 
     /**
      * 用户昵称
@@ -26,12 +18,17 @@ public class UserQueryRequest extends PageRequest implements Serializable {
     private String userAccount;
 
     /**
-     * 简介
+     * 用户头像
+     */
+    private String userAvatar;
+
+    /**
+     * 用户简介
      */
     private String userProfile;
 
     /**
-     * 用户角色：user/admin/ban
+     * 用户角色: user, admin
      */
     private String userRole;
