+++ src/main/java/com/kyle/kyleaicodegen/model/dto/user/UserQueryRequest.java
@@ -1,11 +1,14 @@
-package com.kyle.kyleaicodegen.model.dto;
+package com.kyle.kyleaicodegen.model.dto.user;
 
+import com.kyle.kyleaicodegen.common.PageRequest;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 
 import java.io.Serializable;
 
+@EqualsAndHashCode(callSuper = true)
 @Data
-public class UserUpdateRequest implements Serializable {
+public class UserQueryRequest extends PageRequest implements Serializable {
 
     /**
      * id
@@ -18,9 +21,9 @@ public class UserUpdateRequest implements Serializable {
     private String userName;
 
     /**
-     * 用户头像
+     * 账号
      */
-    private String userAvatar;
+    private String userAccount;
 
     /**
      * 简介
@@ -28,7 +31,7 @@ public class UserUpdateRequest implements Serializable {
     private String userProfile;
 
     /**
-     * 用户角色：user/admin
+     * 用户角色：user/admin/ban
      */
     private String userRole;
